Move report plotting to analysis_plot_utils
